rename query_field to query_bar_field
