+++ x-pack/plugins/security_solution/public/detection_engine/rule_creation_ui/components/query_bar_field/constants.ts
