+++ x-pack/plugins/security_solution/public/detection_engine/rule_creation_ui/components/query_bar_field/query_field.test.tsx
@@ -7,7 +7,7 @@
 
 import React from 'react';
 
-import { QueryField } from '.';
+import { QueryBarField } from '.';
 import {
   TestProviders,
   useFormFieldMock,
@@ -74,7 +74,7 @@ describe('QueryBarDefineRule', () => {
     const { getByTestId } = render(
       <TestProviders>
         <Router history={mockHistory}>
-          <QueryField
+          <QueryBarField
             isLoading={false}
             indexPattern={{ fields: [], title: 'title' }}
             onCloseTimelineSearch={jest.fn()}
@@ -96,7 +96,7 @@ describe('QueryBarDefineRule', () => {
       const { queryByTestId } = render(
         <TestProviders>
           <Router history={mockHistory}>
-            <QueryField
+            <QueryBarField
               isLoading={false}
               indexPattern={{ fields: [], title: 'title' }}
               onCloseTimelineSearch={jest.fn()}
@@ -121,7 +121,7 @@ describe('QueryBarDefineRule', () => {
     const { getByTestId } = render(
       <TestProviders>
         <Router history={mockHistory}>
-          <QueryField
+          <QueryBarField
             isLoading={false}
             indexPattern={{ fields: [], title: 'title' }}
             onCloseTimelineSearch={jest.fn()}
