+++ x-pack/plugins/security_solution/public/detection_engine/rule_creation_ui/validators/query_required_validator_factory.ts
@@ -10,7 +10,7 @@ import { isEmpty } from 'lodash';
 import type { RuleType } from '@kbn/securitysolution-rules';
 import type { FormData, ValidationFunc } from '../../../shared_imports';
 import { isEqlRule, isEsqlRule } from '../../../../common/detection_engine/utils';
-import type { FieldValueQueryBar } from '../components/query_field';
+import type { FieldValueQueryBar } from '../components/query_bar_field';
 
 export function queryRequiredValidatorFactory(
   ruleType: RuleType
