+++ x-pack/plugins/security_solution/public/detection_engine/rule_management/components/rule_details/three_way_diff/final_edit/fields/esql_query/esql_query_edit_form.tsx
@@ -8,7 +8,7 @@
 import React from 'react';
 import type { FormData, FormSchema } from '../../../../../../../../shared_imports';
 import { RuleFieldEditFormWrapper } from '../rule_field_edit_form_wrapper';
-import type { FieldValueQueryBar } from '../../../../../../../rule_creation_ui/components/query_field';
+import type { FieldValueQueryBar } from '../../../../../../../rule_creation_ui/components/query_bar_field';
 import {
   type DiffableRule,
   QueryLanguageEnum,
